- Don’t compile http server stuff. - Cleaned up buttons in gallery view. - Fixed memory leak nav items. - Don’t record layer opacity changes when moving sliders in layers view. - Don’t use direct brushes. - Fixed various warnings.

+++ klodder/Classes/View_BrushSelect.mm
@@ -150,7 +150,7 @@ BrushItem::BrushItem(ToolType type, Brush_Pattern* pattern)
 	
 	for (size_t i = 0; i < itemList.size(); ++i)
 		if (itemList[i].mPatternId == patternId)
-			index = i;
+			index = (int)i;
 	
 	if (index == -1)
 		return;
